feat(tools): Make tool installation common across all containers
Moves the list of tools to apt-get install in the various containers
into the top-level Makefile so the tools available to every AI agent
are comment, easier to maintain.

There are defaults in the Dockerfiles to make things work if you still
want to build the containers with a `build` command, but really building
is now something that needs to happen from the top level of the repo
with `make` to be certain it will work correctly.