feat(shell): Common shell configuration
Bootstraps a way to include common shell customizations into the
containers. You are not usually dropping to a shell in these containers,
but when you do it's nice to have them not suck.

The Makefile updates to copy the common/bash_aliases file in to the
sub-directories is necessary otherwise the container build tools fail
on the COPY call complaining about container escape issues.